+++ back/src/controllers/categories/CreateCategory.go
@@ -2,9 +2,10 @@ package controllers
 
 import (
 	"encoding/json"
-	"github.com/KrispyTech/airneis/config"
+
 	"github.com/KrispyTech/airneis/lib/shared/constants"
-	model "github.com/KrispyTech/airneis/src/db/models"
+	"github.com/KrispyTech/airneis/pkg/config"
+	model "github.com/KrispyTech/airneis/src/models"
 	"github.com/gofiber/fiber/v2"
 )
